change acceleration to a Decimal

+++ data/models.py
@@ -1,9 +1,9 @@
 from django.db import models
 
-# Create your models here.
+
 class Acceleration(models.Model):
     device_id = models.CharField(max_length=255)
-    acceleration = models.IntegerField()
+    acceleration = models.DecimalField()
     timestamp = models.DateTimeField()
     create_time = models.DateTimeField(auto_now_add=True)
     modify_time = models.DateTimeField(auto_now=True)
